backend/requirements.txt
psycopg2-binary~=2.9.10
fastapi~=0.115.3
uvicorn~=0.32.0
minio~=7.2.0
python-multipart~=0.0.9